PCHR-3084: Move requirejs tasks in a separate module
